--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,13 @@
+*.swp
+*.consign
+*.d
+.sl64_gcc4*
+*.root
+*.pdf
+*.png
+*.gif
+*.jpg
+.sl64_gcc447
+temp_gccflags.c
+*.o
+*.so