--- a/.gitmodules
+++ b/.gitmodules
@@ -1,0 +1,3 @@
+[submodule "star-cmake"]
+	path = star-cmake
+	url = https://github.com/star-bnl/star-cmake.git