--- a/StPicoDstMaker/StPicoEnumerations.h
+++ b/StPicoDstMaker/StPicoEnumerations.h
@@ -1,0 +1,7 @@
+#ifndef StPicoEnumerations_h
+#define StPicoEnumerations_h
+
+enum StPicoVtx {PicoVtxAuAu200};
+enum { nTrk = 50000};
+
+#endif