commit 963ca8583babc1cfc405faff323536d1b2cb2fd8
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Thu Aug 11 00:32:38 2016 -0400

    Clean up

commit cb04e1bbd88e93cbae8258411c18f9e56404acb3
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Thu Aug 11 00:21:02 2016 -0400

    Move containers into a separate directory. Correct include files accordingly

commit d6ebff14779b77409d893685df3ac8e6fa16a4bc
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 21:01:35 2016 -0400

    User pre-increment instead of post-increment

commit 71791a75d1cb9d79722d08d1af1c04cdffb4f33f
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 21:01:27 2016 -0400

    Remove legacy commented code

commit e8970ebd71cadf130cc8b1e1269bd68cfb1365bf
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 19:59:27 2016 -0400

    Remove unused methods

commit ba8bd25e11af3f8aa00b7e81843e4da91837e1cf
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 19:59:03 2016 -0400

    1)Emphasize pointer type (this is C++!) 2) Use universal constannes expression

commit 1094a97dbfcf93b20b93e9bd5b18b32207faf1b0
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 19:11:09 2016 -0400

    Remove unused headers and forward declarations. Revisit order of includes

commit c8a6104725f949754bf6c06ff59e0f4900ee66b4
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 18:23:16 2016 -0400

    Fix formatting - astyle

commit e91b8212fc09a32761c2f261e90df9acd9ffcb0e
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 18:05:57 2016 -0400

    Remove member prefix from non-member variable

commit 735990628af4279073db9a20a870a842968ee796
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 17:42:36 2016 -0400

    Don't apply any event or track cuts

commit e2c4f655216360ffa38a98d2356a2312a3f2d828
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Wed Aug 10 13:26:02 2016 -0400

    Make vertex selection mode a separate mode

commit 05b9f24dd10301ef74f40cc1221e0e1bafc3bc92
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Fri Aug 5 15:26:44 2016 -0400

    Replace track parameters with gMom and Origin. Remove track parameters errors

commit fb8003ba658fae3f59aa75339d84bc14ef1e2dc6
Author: Mustafa Mustafa <mmustafa@lbl.gov>
Date:   Thu Aug 4 21:54:34 2016 -0400

    Remove unused macro

commit 705b6d8ef225061ebc7baa6246c4a89e5c996522
Author: RongrongMa <marr@bnl.gov>
Date:   Thu Aug 4 12:00:44 2016 -0400

    Change the cut for |J2-J3| to be 1023 for Run15

commit cc18bc358f367a29576c627bea6915a1f6d7ea07
Author: RongrongMa <marr@bnl.gov>
Date:   Mon Aug 1 17:59:28 2016 -0400

    The lower threshold of MTD QT channel was changed from 100 to 80
    since run 16045067

commit 6e5895084f7276b804c5be6dcda0d3ff4b1b9d20
Author: RongrongMa <marr@bnl.gov>
Date:   Thu Jul 21 11:45:47 2016 -0400

    Calculate year within the local domain of function initMtd() and fillMtdTrigger()

commit 74cbd548498932f78c4aa6d4e28c488be9f036f9
Author: RongrongMa <marr@bnl.gov>
Date:   Tue Jul 19 14:38:52 2016 -0400

    Remove unnecessary header

commit 45c860496b014e8e0b7c36cd563d9ecc1298da17
Author: RongrongMa <marr@bnl.gov>
Date:   Tue Jul 19 14:19:57 2016 -0400

    1) Add slewing correction to QT information.
    2) Accomandate 4 additional QT boards installed for Run16
    3) Add MtdHitIndex to MTD pidTraits

commit daab41fe321298340cdaa1b59d97847ba5da32e6
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Mon Jul 18 15:10:48 2016 -0400

    Added missing header for numeric_limits

commit 4d32e5bd85daf0b17500d24028a796ce94746e12
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri Jul 15 01:50:48 2016 -0400

    Took out calls to ROOT's ClassImp macro

    For all practical reasons the inclusion of ClassImp macro does not seem to be
    necessary in recent versions of ROOT. Static objects of TGenericClassInfo are
    instantiated in dictionary files generated by rootcint which is triggered by
    the presence of the ClassDef macro. The registration of class implementation for
    THtml documentation also appears to be irrelevant for STAR.

commit 2ea71f617671c66d4b91cf0be93bf477676395fb
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Tue May 31 15:57:39 2016 -0400

    StPicoTrack: Added interface to check if track is associated with a primary vertex

commit e5a280abc327cf76c39da75ed4a3ab170492fe26
Author: RongrongMa <marr@bnl.gov>
Date:   Thu Jul 14 12:01:18 2016 -0400

    1) Include full path for out-of-directory header files
    2) Add data member mShouldHaveRejectEvent to indicate event status
    during two-pass iteration reconstruction.

commit f8db5782dd0aaf60615eae060d9a443949f47c26
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Jul 13 15:12:22 2016 -0700

    Add track index access method

commit fe0450c2b373e2dd2747652115402e33cb5f4997
Author: RongrongMa <marr@bnl.gov>
Date:   Wed Jul 13 16:53:42 2016 -0400

    Added following data members:
    1. mVpdTacSum - collision start time measured by VPD
    2. mTHUBtime[] - collision trigger time recorded in THUB

commit b367f5657caa6d3fc6a0c3e80108b4fec53e1982
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri May 20 14:45:29 2016 -0700

    Simplify ::openRead syntax

commit cd85c1dddf557376c43e8fa4fabad9219aed4470
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri May 20 14:45:09 2016 -0700

    Simplify work in constructors

commit d17df84caf74c90273558a529237d76b678ed33c
Author: Ji Xu <xuj@rcas6008.rcf.bnl.gov>
Date:   Fri May 20 17:18:25 2016 -0400

    Add refmult3/remult4 to picoEvent

commit a7fa7f7b6ae43b6c3d8fa56955b1f487a1c0d95f
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 14:44:15 2016 -0400

    Added how to build library instructions

commit c25beb8bd1234c6ddcaa0b3842ac6c1b967c6322
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 14:38:42 2016 -0400

    cmake: Added paths to header files in standard STAR environment

    This is for those who don't want to provide by the paths in the command line by
    hand

commit 781ab724a759d1efbfd07d356c009647bda46277
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 14:37:22 2016 -0400

    cmake: Updated example with command to build the library

commit 831e84d86be8732673982dca8a898f0fef47b1b3
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 00:12:56 2016 -0400

    travis-ci: Significantly reworked steering script

    The build commands are the same as in the other repositories in star-bnl

commit fe12094a36879f2145cf07b6bfd1787f460d1904
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 00:09:43 2016 -0400

    README: Initial commit showing build status on travis-ci

commit 238a329679f5b40a5bf913ce4954b84a826ab2e5
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 00:04:36 2016 -0400

    cmake: Remove deleted file StPicoV0.cxx from library

commit 887136c9355da4de2a6862ed1cfa250c12a1f72d
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 00:03:28 2016 -0400

    cmake: Correct location for math_ and phys_constants.h

commit 9d485c5f83afd5d747776263449a31d8e57cd2d7
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Fri May 20 00:02:17 2016 -0400

    cmake: Replaced local cmake/ with common external from star-bnl repo

commit 50204c478d34618b7bcf7072bc7a5795d1b14b95
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Thu May 19 19:30:30 2016 -0400

    Delete objects owned by StPicoDstMaker

commit 3033f6c79cc6277669217037ca90833a0a915430
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Thu May 19 11:45:44 2016 -0400

    StPicoDstMaker: [Cosmetic] Got rid of obscuring macro

commit a33dffd9115d08a97f387f1eebc3116de49aa354
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Thu May 19 14:36:34 2016 -0400

    StPicoDstMaker: Remove unnecessary data members

commit 976ebff6612cc5b095f99b797bcff1a219d20a80
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Thu May 19 14:31:37 2016 -0400

    StPicoDstMaker: Do not ask rootcint to generate class streamer

    STAR makers are not persistent objects

commit ad72866122304b635bd3ac05070132481a7cddfd
Author: Dmitri Smirnov <d.s@plexoos.com>
Date:   Thu May 19 12:09:28 2016 -0400

    StPicoDstMaker: Do not expose namespaces unnecessarily

commit 2543251319115687846a9dc2d6588e52890c22a9
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri May 13 10:39:29 2016 -0700

    Slightly faster version

commit 0bf5477d9e97965d86bedb2ee18e16d1634b901a
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri May 13 10:31:02 2016 -0700

    Change nHitsMapHft to hftHitsMap

    Previous name is confusing

commit 50b018439ca21856a3ed9a4310bf3da045c84cf1
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu May 12 07:21:22 2016 -0700

    Use unsigned int for bit manipulation

    Variable bits encode patterns, should always be unsigned int

commit 7237e6d45c7a785f1b021f74ea1ba7008ba02214
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Sun May 1 07:48:46 2016 -0700

    add missing header

commit fbcc6343b81841023b2b0e14ef17303218da39c8
Author: Guannan Xie <GuannanXie@lbl.gov>
Date:   Fri Apr 29 19:14:19 2016 -0400

    return value for the function  runId()

commit 9b05a7bf5e8c12834c362afb6c0e14223e2cc142
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri Mar 25 13:13:00 2016 -0700

    add ZDC channels readout

commit af7be70e8bd39b6a1a3742ea2809ca94030e3c63
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Fri Mar 25 09:38:47 2016 -0700

    Remove unsed method

commit a087cccf7ad9e6f4bc8a7356a42740127617cab4
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 16:40:20 2016 -0700

    remove trailing whitespace

commit be928c46f5fdad940fa8cd2ae4a1f2e583f67258
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 16:14:32 2016 -0700

    Remove StPicoDstMaker friendship; not needed and not allowed

commit 6745ebaa4f6d42d3614f3e3cef883db905434c00
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 16:01:19 2016 -0700

    Remove SpaceCharge

commit 201598bd2e935fa0a8aee46579dc2f87ff24a79a
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 16:00:15 2016 -0700

    Remove ZDC and VPD channels readouts

commit 18ffc84302d236b7ac04f4e645e23d8e611393a5
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 11:09:37 2016 -0700

    remove centrality definitions

commit 5babe609127907ecc6428cfc4c0b2d075fb592e5
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 11:09:01 2016 -0700

    remove trigger constants remains

commit 05e6f7442841d61f049e9d8a6fdf4396a53d4c70
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:34:56 2016 -0700

    Remove basic types max definitions; they are no longer used

commit 19705bcdbcdb0396757dbaeebff62674fd7232c7
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:32:18 2016 -0700

    Remove trigger constants; they are not used anymore

commit ee05d1d484f57c02b2ec15a8ef7ac127a7942990
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:29:46 2016 -0700

    Remove unnecessary comments

commit d45aa7bab4a29b0497f1c6efe3106376c37a964e
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:25:55 2016 -0700

    Remove JP thresholds; they are never even filled

commit 8da18e8baa7c953a5ea24607d528ea5fd1d8cb0a
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:24:32 2016 -0700

    remove friends; no friends allowed

commit 4fa5f1e688c98ea8946b4b551f8ddb1ab1402d6d
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 10:13:49 2016 -0700

    remove unnecessary includes

commit 3512e4fd02e7b022014e009e458af1be2666f929
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Thu Mar 24 09:46:47 2016 -0700

    streamline the data members. Initialize variables in initializer list. Delegate construction.

commit 10bdca19017e2fd69861a540e989d44b916119d6
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 23 15:36:29 2016 -0700

    remove trigger word scheme and directly save trigger ids

commit 40f2610ccf190b946c2cf45e85b8d9d201310ef5
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 23 13:20:29 2016 -0700

    define methods outside the header

commit ba4aafee01fa4b9758ac1004052a92f8590e7cea
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 16 13:50:29 2016 -0700

    StPicoEvent constructor doesn't need Q anymore

commit 89416b24e8cee56e2cb932dba4cdd702fb2ec097
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 16 13:15:33 2016 -0700

    Remove all phi weight and flow flag related codes

commit 05cefccd5a67cf79b3649a1aea2ed6e22238c7d1
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 16 13:04:07 2016 -0700

    Fix bug: 1) uninitialized variables 2) return one track projection is outside BTOW acceptance

commit ed046bf256231b8b6e502655154f18ee2fb218d2
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 15:01:40 2016 -0700

    Completely remove V0 related codes

commit 7ff38a69e9f8f03cab1a1af40620389321ad8f24
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:21:33 2016 -0700

    Irnon constructors

commit 82f9c5f9f09ff029a38c122a206e2881756f4088
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:20:07 2016 -0700

    define methods outside the headeR

commit f6761333df3c1d83d44dd6086eb352a8493ad79d
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:19:35 2016 -0700

    Ironing constructors

commit 277808ed24b41ffed7876fa0c1a84344fe00772d
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:18:57 2016 -0700

    define methods outside the header

commit 99eb0f5837379f3bc91c3369abcffa03f3a57357
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:17:16 2016 -0700

    remove unused header

commit 9b2300e1181aea5b7d8dafdbf9405cb7de33f058
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:17:00 2016 -0700

    No friends allowed!

commit ac382b98a4c2b3e632709305c8ac535452cc2bda
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:16:29 2016 -0700

    TObject::Clear definition is not needed. Iron constructors

commit d763f28c2da483318641e1d0e2e04d47d58745f4
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:15:32 2016 -0700

    define methods outside the header

commit a0fbfc925a83e04d759a0a2a9d93ab0dda484052
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:14:17 2016 -0700

    delegate the construction

commit a2455096893cefcf6cf592639ffc613dcf7cf1eb
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:13:06 2016 -0700

    the values are already set to zero in the initializer list

commit db692478afa6029e3577d7074e3ec9f2f75f2ce5
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:12:41 2016 -0700

    the values are already set to zero in the initializer list

commit 2a0f734bf28a16e74c3102d78dcb8698423d570d
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:07:54 2016 -0700

    ironing

commit fe949e5b039d31853f410fb2171be1ecda6bea32
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:07:10 2016 -0700

    header file not used

commit 0300a10c5f11b12916cd6cb14e9d082516d1c2cd
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:06:54 2016 -0700

    stream line constructors

commit 9115cd76a4f8944ba75c72546bc1b5e9aa453c2f
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 14:05:51 2016 -0700

    define methods outside the header file

commit 6960f9905c3ed54de392926256fd7d6f5533972b
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:51:14 2016 -0700

    methods should defined outside the header

commit fa61e127024422039dcdf6dad9b74519e1c6fd22
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:43:50 2016 -0700

    using namespace std in a header file is not allowed

commit 3df612605e534a848fb31caa65e6575e31372532
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:31:35 2016 -0700

    Use intializer list for intializing variables. Iron constructors.

commit 16be0e7f3b27c5f0ccaf5cc23a570dc08af61bf9
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:30:13 2016 -0700

    as per guidelines, methods should be defined out side of the header

commit 6f5b7c1ccd53db41a9eba9cde24a86ef32f03910
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:26:36 2016 -0700

    No friends allowed! Actually, not needed

commit c62e1ffaeec87c2635e0f57152ac2422c2cf7118
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:26:08 2016 -0700

    iron constructors. Remove definition of TObject::Clear(); not needed

commit 62a1b37ba7d58a8af2054d83f7f584770b523c4a
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:19:56 2016 -0700

    as per the guidelines, methods should be define outside the header

commit dfd9f449b1864a567c48fba32b2725dca68d9ea1
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:18:01 2016 -0700

    No friends allowed! (not even needed)

commit 2f09f19f8b64fadf81d262fe3a368a9bac91ca66
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:17:01 2016 -0700

    iron the constructors and their intializer lists

commit 2759306af274727b698259beafc35377133abf45
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:14:21 2016 -0700

    Overriding TObject::Clear method is not needed here. The constructors are called everytime and they reset the variables in their intializer lists

commit 4b799e47abc3e304b9c6ca1863016aaf6e65804d
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 09:10:40 2016 -0700

    as per the guidelines, methods should be defined outside the class header.

commit 35b454d70bdbc9d57610709674800cb5c7812336
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Mon Mar 14 08:34:36 2016 -0700

    protection against going beyond array limit (added by Rongrong)

commit f42efaf9b67122a5bd6e703f180ec93cb3477287
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 9 14:17:41 2016 -0800

    This class doesn't need to be a friend of StPicoDst

commit c61b1c3d8feca6179f720d56361524cd39054543
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 9 13:49:05 2016 -0800

    review constructor signature; remove flow weight and flow Flag

commit 72a41b1fe3032736b8e02a128934fdac4489dca6
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 9 13:38:48 2016 -0800

    remove unnecessary includes. Avoid transitive includes

commit 2e413e8a8fbb7075fa570e81f584bd777173ed76
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 9 12:58:09 2016 -0800

    as per the guidelines, methods should be defined outside the class header

commit fe80dce57702760b2516e362e9aecf27fe9770c6
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Wed Mar 9 11:06:26 2016 -0800

    stream line constructors and member variable defaults

commit 4935ff5c99439863ddb19c09aba50167dabd270c
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Tue Mar 8 13:48:58 2016 -0800

    fix #define guard format to be consistent with the guidelines

commit 2206c2ae5dabc2276358efbea830497bba4434dd
Author: Mustafa Mustafa <mstftsm@gmail.com>
Date:   Tue Mar 8 13:37:12 2016 -0800

    remove excute privilage from all users; not needed for a macro

commit a007a99c4a3eed7b0e1ad7275bf9730da2d7b93f
…